Create CrossProduct.py
The program is given two matrices n×k and k×m as input. In the first line, the natural values ​​n, k, and m are given to the program, respectively, with a space between them. In the following lines, the values ​​of the two matrices will be given to the program. In the output, the program should output the final n×m matrix obtained by multiplying the first matrix by the second matrix.